GeneralizedMultiparticleMie

Documentation for GeneralizedMultiparticleMie.

GeneralizedMultiparticleMie.associated_legendre

GeneralizedMultiparticleMie.associated_legendre_array

GeneralizedMultiparticleMie.expand_E_cluster

GeneralizedMultiparticleMie.factorial

GeneralizedMultiparticleMie.gaunt_a

GeneralizedMultiparticleMie.gaunt_b

GeneralizedMultiparticleMie.init_vswf_cache

GeneralizedMultiparticleMie.pi_func

GeneralizedMultiparticleMie.ricatti_hn1

GeneralizedMultiparticleMie.ricatti_hn1_deriv

GeneralizedMultiparticleMie.ricatti_hn2

GeneralizedMultiparticleMie.ricatti_hn2_deriv

GeneralizedMultiparticleMie.ricatti_jn

GeneralizedMultiparticleMie.ricatti_jn_deriv

GeneralizedMultiparticleMie.ricatti_yn

GeneralizedMultiparticleMie.ricatti_yn_deriv

GeneralizedMultiparticleMie.spherical_hn1

GeneralizedMultiparticleMie.spherical_hn1_deriv

GeneralizedMultiparticleMie.spherical_hn2

GeneralizedMultiparticleMie.spherical_hn2_deriv

GeneralizedMultiparticleMie.spherical_jn

GeneralizedMultiparticleMie.spherical_jn_deriv

GeneralizedMultiparticleMie.spherical_yn

GeneralizedMultiparticleMie.spherical_yn_deriv

GeneralizedMultiparticleMie.tau_func

GeneralizedMultiparticleMie.vswf_electric

GeneralizedMultiparticleMie.vswf_magnetic

GeneralizedMultiparticleMie.wigner_3j

GeneralizedMultiparticleMie.wigner_d

GeneralizedMultiparticleMie.wigner_d_recursion

GeneralizedMultiparticleMie.associated_legendre — Method

Associated Legendre function without the Condon-Shotley phase.

Arblib's definition includes the Condon-Shotley phase, so we need to multiply the results by $(-1)^m$.

GeneralizedMultiparticleMie.associated_legendre_array — Method

Calculate the associated Legendre function for all $0\leq n\leq n_{\max}$ and $-n\leq m\leq n$, and return the results as a vector.

Arblib's Legendre function definition includes the Condon-Shotley phase, so we need to multiply the results by $(-1)^m$.

GeneralizedMultiparticleMie.expand_E_cluster — Method

Expand the electric field.

GeneralizedMultiparticleMie.factorial — Method

Calculate factorials using the Gamma function.

GeneralizedMultiparticleMie.gaunt_a — Method

The Gaunt a-coefficient defined by Gaunt (1929):

\[\begin{aligned} a(m, n, \mu, \nu, p)=&(-1)^{m+\mu}(2 p+1)\left[\frac{(n+m) !(\nu+\mu) !(p-m-\mu) !}{(n-m) !(\nu-\mu) !(p+m+\mu) !}\right]^{1 / 2} \\ & \times\left(\begin{array}{ccc} n & \nu & p \\ 0 & 0 & 0 \end{array}\right)\left(\begin{array}{ccc} n & \nu & p \\ m & \mu & -m-\mu \end{array}\right) \end{aligned}\]

References:

Gaunt, J.A., 1929. IV. The triplets of helium. Philosophical Transactions of the Royal Society of London. Series A, Containing Papers of a Mathematical or Physical Character 228, 151–196.

GeneralizedMultiparticleMie.gaunt_b — Method

\[\begin{aligned} b(m, n, \mu, \nu, p)=&(-1)^{m+\mu}(2 p+3)\left[\frac{(n+m) !(\nu+\mu) !(p-m-\mu) !}{(n-m) !(\nu-\mu) !(p+m+\mu) !}\right]^{1 / 2} \\ & \times\left(\begin{array}{ccc} n & \nu & p \\ 0 & 0 & 0 \end{array}\right)\left(\begin{array}{ccc} n & \nu & p+1 \\ m & \mu & -m-\mu \end{array}\right) \end{aligned}\]

GeneralizedMultiparticleMie.init_vswf_cache — Method

Precompute required VSWF coefficients.

GeneralizedMultiparticleMie.pi_func — Method

Calculate $\pi_n^m(\theta)$ defined as

\[\pi_n^m(\theta)=\frac{m}{\sin\theta}P_n^m(\cos\theta)\]

GeneralizedMultiparticleMie.ricatti_hn1 — Method

Riccati Hankel function of the first kind.

GeneralizedMultiparticleMie.ricatti_hn1_deriv — Method

First-order derivative of Riccati Hankel function of the first kind.

GeneralizedMultiparticleMie.ricatti_hn2 — Method

Riccati Hankel function of the second kind.

GeneralizedMultiparticleMie.ricatti_hn2_deriv — Method

First-order derivative of Riccati Hankel function of the second kind.

GeneralizedMultiparticleMie.ricatti_jn — Method

Riccati Bessel function of the first kind.

GeneralizedMultiparticleMie.ricatti_jn_deriv — Method

First-order derivative of Riccati Bessel function of the first kind.

GeneralizedMultiparticleMie.ricatti_yn — Method

Riccati Bessel function of the second kind.

Note that in miepy, the author used -z⋅y(z) instead of z⋅y(z)

GeneralizedMultiparticleMie.ricatti_yn_deriv — Method

First-order derivative of Riccati Bessel function of the second kind.

GeneralizedMultiparticleMie.spherical_hn1 — Method

Spherical Hankel function of the first kind.

GeneralizedMultiparticleMie.spherical_hn1_deriv — Method

First-order derivative of spherical Hankel function of the first kind.

GeneralizedMultiparticleMie.spherical_hn2 — Method

Spherical Hankel function of the second kind.

GeneralizedMultiparticleMie.spherical_hn2_deriv — Method

First-order derivative of spherical Hankel function of the second kind.

GeneralizedMultiparticleMie.spherical_jn — Method

Spherical Bessel function of the first kind.

GeneralizedMultiparticleMie.spherical_jn_deriv — Method

First-order derivative of spherical Bessel function of the first kind.

GeneralizedMultiparticleMie.spherical_yn — Method

Spherical Bessel function of the second kind.

GeneralizedMultiparticleMie.spherical_yn_deriv — Method

First-order derivative of spherical Bessel function of the second kind.

GeneralizedMultiparticleMie.tau_func — Method

Calculate $\tau_n^m(\theta)$ defined as

\[\tau_n^m(\theta)=\frac{\mathrm{d}}{\mathrm{d}\theta}P_n^m(\cos\theta)\]

GeneralizedMultiparticleMie.vswf_electric — Method

Vector spherical wave function, electric (TM) modes.

\[\begin{aligned} \mathbf{N}_{m n}^{(i)}(\rho, \theta, \phi)=&\left\{\hat{\mathbf{e}}_{r} n(n+1) P_{n}^{m}(\cos \theta) \frac{z^{(i)}_{n}(\rho)}{\rho}\right.\\ &\left.+\left[\hat{\mathbf{e}}_{\theta} \tau_{m n}(\cos \theta)+\hat{\mathbf{e}}_{\phi} \mathrm{i} \pi_{m n}(\cos \theta)\right] \frac{(\rho\cdot z_n^{(i)}(\rho))^{\prime}}{\rho}\right\} \exp (\mathrm{i} m \phi) \end{aligned}\]

where $i=1,2,3,4$ relates to incident, interior, outgoing and ingoing VSWF.

GeneralizedMultiparticleMie.vswf_magnetic — Method

Vector spherical wave function, magnetic (TE) modes.

\[\mathbf{M}_{m n}^{(i)}(\rho, \theta, \phi)=\left[\hat{\mathbf{e}}_{\theta} \mathrm{i} \pi_{m n}(\cos \theta)-\hat{\mathbf{e}}_{\phi} \tau_{m n}(\cos \theta)\right] z_{n}^{(i)}(\rho) \exp (\mathrm{i} m \phi)\]

where $i=1,2,3,4$ relates to incident, interior, outgoing and ingoing VSWF.

GeneralizedMultiparticleMie.wigner_3j — Function

Wigner 3-j symbols.

GeneralizedMultiparticleMie.wigner_d — Method

Calculate Wigner (small) d-function $d_{mn}^s(\theta)$ for a single $(m, n, s)$ combination, using Eq. (B.1) of Mishchenko et al. (2002).

\[\begin{aligned} d_{m n}^{s}(\vartheta)=& \sqrt{(s+m) !(s-m) !(s+n) !(s-n) !} \\ & \times \sum_{k=\max(0,m-n)}^{\min(s + m, s - n)}(-1)^{k} \frac{\left(\cos \frac{1}{2} \vartheta\right)^{2 s-2 k+m-n}\left(\sin \frac{1}{2} \vartheta\right)^{2 k-m+n}}{k !(s+m-k) !(s-n-k) !(n-m+k) !} \end{aligned}\]

GeneralizedMultiparticleMie.wigner_d_recursion — Method

Calculate Wigner (small) d-function $d_{mn}^s(\theta)$ for $s\in[s_{\min}=\max(|m|, |n|),s_{\max}]$ (and also its derivative) via upward recursion, using Eq. (B.22) of Mishchenko et al. (2002).

\[\begin{aligned} d_{m n}^{s+1}(\vartheta)=& \frac{1}{s \sqrt{(s+1)^{2}-m^{2}} \sqrt{(s+1)^{2}-n^{2}}}\left\{(2 s+1)[s(s+1) x-m n] d_{m n}^{s}(\vartheta)\right.\\ &\left.-(s+1) \sqrt{s^{2}-m^{2}} \sqrt{s^{2}-n^{2}} d_{m n}^{s-1}(\vartheta)\right\}, \quad s \geq s_{\min } \end{aligned}\]

The initial terms are given by Eq. (B.23) and Eq. (B.24).

\[\begin{array}{l} d_{m n}^{s_{\min }-1}(\vartheta)=0 \\ d_{m n}^{s_{\min }}(\vartheta)=\xi_{m n} 2^{-s_{\min }}\left[\frac{\left(2 s_{\min }\right) !}{(|m-n|) !(|m+n|) !}\right]^{1 / 2}(1-x)^{|m-n| / 2}(1+x)^{|m+n| / 2} \end{array}\]

where

\[\xi_{m n}=\left\{\begin{array}{ll} 1 & \text { for } n \geq m \\ (-1)^{m-n} & \text { for } n<m \end{array}\right.\]