GeneralizedMultiparticleMie

Documentation for GeneralizedMultiparticleMie.

GeneralizedMultiparticleMie.a_func

GeneralizedMultiparticleMie.associated_legendre

GeneralizedMultiparticleMie.associated_legendre_array

GeneralizedMultiparticleMie.b_func

GeneralizedMultiparticleMie.factorial

GeneralizedMultiparticleMie.pi_func

GeneralizedMultiparticleMie.ricatti_hn1

GeneralizedMultiparticleMie.ricatti_hn1_deriv

GeneralizedMultiparticleMie.ricatti_hn2

GeneralizedMultiparticleMie.ricatti_hn2_deriv

GeneralizedMultiparticleMie.ricatti_jn

GeneralizedMultiparticleMie.ricatti_jn_deriv

GeneralizedMultiparticleMie.ricatti_yn

GeneralizedMultiparticleMie.ricatti_yn_deriv

GeneralizedMultiparticleMie.spherical_hn1

GeneralizedMultiparticleMie.spherical_hn1_deriv

GeneralizedMultiparticleMie.spherical_hn2

GeneralizedMultiparticleMie.spherical_hn2_deriv

GeneralizedMultiparticleMie.spherical_jn

GeneralizedMultiparticleMie.spherical_jn_deriv

GeneralizedMultiparticleMie.spherical_yn

GeneralizedMultiparticleMie.spherical_yn_deriv

GeneralizedMultiparticleMie.tau_func

GeneralizedMultiparticleMie.wigner_3j

GeneralizedMultiparticleMie.a_func — Method

a-function.

GeneralizedMultiparticleMie.associated_legendre — Method

Associated Legendre function without the Condon-Shotley phase.

Arblib's definition includes the Condon-Shotley phase, so we need to multiply the results by $(-1)^m$.

GeneralizedMultiparticleMie.associated_legendre_array — Method

Calculate the associated Legendre function for all $0\leq n\leq n_{\max}$ and $-n\leq m\leq n$, and return the results as a vector.

Arblib's Legendre function definition includes the Condon-Shotley phase, so we need to multiply the results by $(-1)^m$.

GeneralizedMultiparticleMie.b_func — Method

b-function.

GeneralizedMultiparticleMie.factorial — Method

Calculate factorials using the Gamma function.

GeneralizedMultiparticleMie.pi_func — Method

Calculate $\pi_n^m(\theta)$ defined as

\[\pi_n^m(\theta)=\frac{m}{\sin\theta}P_n^m(\cos\theta)\]

GeneralizedMultiparticleMie.ricatti_hn1 — Method

Riccati Hankel function of the first kind.

GeneralizedMultiparticleMie.ricatti_hn1_deriv — Method

First-order derivative of Riccati Hankel function of the first kind.

GeneralizedMultiparticleMie.ricatti_hn2 — Method

Riccati Hankel function of the second kind.

GeneralizedMultiparticleMie.ricatti_hn2_deriv — Method

First-order derivative of Riccati Hankel function of the second kind.

GeneralizedMultiparticleMie.ricatti_jn — Method

Riccati Bessel function of the first kind.

GeneralizedMultiparticleMie.ricatti_jn_deriv — Method

First-order derivative of Riccati Bessel function of the first kind.

GeneralizedMultiparticleMie.ricatti_yn — Method

Riccati Bessel function of the second kind.

Note that in miepy, the author used -z⋅y(z) instead of z⋅y(z)

GeneralizedMultiparticleMie.ricatti_yn_deriv — Method

First-order derivative of Riccati Bessel function of the second kind.

GeneralizedMultiparticleMie.spherical_hn1 — Method

Spherical Hankel function of the first kind.

GeneralizedMultiparticleMie.spherical_hn1_deriv — Method

First-order derivative of spherical Hankel function of the first kind.

GeneralizedMultiparticleMie.spherical_hn2 — Method

Spherical Hankel function of the second kind.

GeneralizedMultiparticleMie.spherical_hn2_deriv — Method

First-order derivative of spherical Hankel function of the second kind.

GeneralizedMultiparticleMie.spherical_jn — Method

Spherical Bessel function of the first kind.

GeneralizedMultiparticleMie.spherical_jn_deriv — Method

First-order derivative of spherical Bessel function of the first kind.

GeneralizedMultiparticleMie.spherical_yn — Method

Spherical Bessel function of the second kind.

GeneralizedMultiparticleMie.spherical_yn_deriv — Method

First-order derivative of spherical Bessel function of the second kind.

GeneralizedMultiparticleMie.tau_func — Method

Calculate $\tau_n^m(\theta)$ defined as

\[\tau_n^m(\theta)=\frac{\mathrm{d}}{\mathrm{d}\theta}P_n^m(\cos\theta)\]

GeneralizedMultiparticleMie.wigner_3j — Function

Wigner 3-j symbols.